requirements.txt
aiogram==3.4.1
SQLAlchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
Pydantic==2.5.0
FastAPI==0.104.1
uvicorn==0.24.0
aiofiles==23.2.1
psycopg2-binary==2.9.9
